Cleanup demo_utils for Yolo models to have common_demo_utils (#26415)
### Problem description
Code duplication for demo utils.

### What's changed
Have a common demo utils that captures common helper fucntions for yolo
and other models.
Current PR removes duplicates for Yolo models such as LoadImages,
load_coco_class_names.

### Checklist

- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16878879814/workflow)
- All passed except ttnn fast runtime off tests (timeout)
- [x] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/runs/16905810772/workflow)
- all passed
- [x] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/runs/16878903694)
- All Passed
- [x] [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/runs/16878887101/workflow)
- all yolos passed
- [x] [Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/16878896005)
- all yolos passed
