+++ models/demos/yolov10x/tests/perf/test_e2e_performant.py
@@ -9,9 +9,9 @@ import torch
 from loguru import logger
 
 import ttnn
+from models.demos.utils.common_demo_utils import get_mesh_mappers
 from models.demos.yolov10x.common import YOLOV10_L1_SMALL_SIZE
 from models.demos.yolov10x.runner.performant_runner import YOLOv10PerformantRunner
-from models.demos.yolov10x.tt.common import get_mesh_mappers
 from models.utility_functions import run_for_wormhole_b0
