+++ models/demos/yolov4/common.py
@@ -128,15 +128,3 @@ def get_model_result(ttnn_output_tensor, resolution, mesh_composer=None):
     result_boxes = torch.cat(result_boxes_list, dim=1)
 
     return [result_boxes.to(torch.float16), result_confs.to(torch.float16)]
-
-
-def get_mesh_mappers(device):
-    if device.get_num_devices() > 1:
-        inputs_mesh_mapper = ttnn.ShardTensorToMesh(device, dim=0)
-        weights_mesh_mapper = None
-        output_mesh_composer = ttnn.ConcatMeshToTensor(device, dim=0)
-    else:
-        inputs_mesh_mapper = None
-        weights_mesh_mapper = None
-        output_mesh_composer = None
-    return inputs_mesh_mapper, weights_mesh_mapper, output_mesh_composer
