+++ models/demos/yolov4/tests/perf/test_e2e_performant.py
@@ -9,7 +9,8 @@ import torch
 from loguru import logger
 
 import ttnn
-from models.demos.yolov4.common import YOLOV4_L1_SMALL_SIZE, get_mesh_mappers
+from models.demos.utils.common_demo_utils import get_mesh_mappers
+from models.demos.yolov4.common import YOLOV4_L1_SMALL_SIZE
 from models.demos.yolov4.runner.performant_runner import YOLOv4PerformantRunner
 from models.perf.perf_utils import prep_perf_report
 from models.utility_functions import run_for_wormhole_b0
