+++ models/demos/yolov6l/demo/demo.py
@@ -8,18 +8,19 @@ import os
 import fiftyone
 import pytest
 import torch
+from loguru import logger
 
 import ttnn
-from models.demos.yolov6l.common import YOLOV6L_L1_SMALL_SIZE, load_torch_model
-from models.demos.yolov6l.demo.demo_utils import *
-from models.demos.yolov6l.demo.demo_utils import (
+from models.demos.utils.common_demo_utils import (
     LoadImages,
+    get_mesh_mappers,
     load_coco_class_names,
     preprocess,
     save_yolo_predictions_by_model,
 )
+from models.demos.yolov6l.common import YOLOV6L_L1_SMALL_SIZE, load_torch_model
+from models.demos.yolov6l.demo.demo_utils import postprocess
 from models.demos.yolov6l.runner.performant_runner import YOLOv6lPerformantRunner
-from models.demos.yolov6l.tt.common import get_mesh_mappers
 from models.utility_functions import disable_persistent_kernel_cache
