+++ models/demos/yolov8s/demo/demo.py
@@ -3,59 +3,23 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
-from datetime import datetime
 
-import cv2
 import pytest
 import torch
 from loguru import logger
 
 import ttnn
-from models.demos.yolov8s.common import YOLOV8S_L1_SMALL_SIZE, load_torch_model
-from models.demos.yolov8s.demo.demo_utils import (
+from models.demos.utils.common_demo_utils import (
     LoadImages,
     get_mesh_mappers,
     load_coco_class_names,
     postprocess,
     preprocess,
+    save_yolo_predictions_by_model,
 )
+from models.demos.yolov8s.common import YOLOV8S_L1_SMALL_SIZE, load_torch_model
 from models.demos.yolov8s.runner.performant_runner import YOLOv8sPerformantRunner
 from models.utility_functions import disable_persistent_kernel_cache
-
-
-def save_yolo_predictions_by_model(result, save_dir, image_path, model_name):
-    model_save_dir = os.path.join(save_dir, model_name)
-    os.makedirs(model_save_dir, exist_ok=True)
-
-    image = cv2.imread(image_path)
-    image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-
-    if model_name == "torch_model":
-        bounding_box_color, label_color = (0, 255, 0), (0, 255, 0)
-    else:
-        bounding_box_color, label_color = (255, 0, 0), (255, 255, 0)
-
-    boxes = result["boxes"]["xyxy"]
-    scores = result["boxes"]["conf"]
-    classes = result["boxes"]["cls"]
-    names = result["names"]
-
-    for box, score, cls in zip(boxes, scores, classes):
-        x1, y1, x2, y2 = map(int, box)
-        label = f"{names[int(cls)]} {score.item():.2f}"
-        cv2.rectangle(image, (x1, y1), (x2, y2), bounding_box_color, 3)
-        cv2.putText(image, label, (x1, y1 - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, label_color, 2)
-
-    image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
-
-    image_base = os.path.splitext(os.path.basename(image_path))[0]
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    output_name = f"{image_base}_prediction_{timestamp}.jpg"
-    output_path = os.path.join(model_save_dir, output_name)
-
-    cv2.imwrite(output_path, image)
-
-    logger.info(f"Predictions saved to {output_path}")
 
 
 def run_yolov8s(
