--- models/demos/yolov8s/demo/demo_utils.py
@@ -1,347 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-import math
-import os
-import time
-from pathlib import Path
-
-import cv2
-import numpy as np
-import requests
-import torch
-import torchvision
-from loguru import logger
-
-import ttnn
-
-
-# Read image using OpenCV from a file path (handles Unicode paths)
-def imread(filename: str, flags: int = cv2.IMREAD_COLOR):
-    return cv2.imdecode(np.fromfile(filename, np.uint8), flags)
-
-
-# Supported image file extensions
-IMG_FORMATS = {"bmp", "dng", "jpeg", "jpg", "mpo", "png", "tif", "tiff", "webp", "pfm", "heic"}
-
-
-# Data loader for image files
-class LoadImages:
-    def __init__(self, path, batch=1, vid_stride=1):
-        files = []
-        for p in sorted(path) if isinstance(path, (list, tuple)) else [path]:
-            a = str(Path(p).absolute())
-            if os.path.isdir(a):
-                for f in os.listdir(a):
-                    if f.lower().endswith((".jpg", ".jpeg", ".png", ".bmp")):
-                        files.append(os.path.join(a, f))
-            elif os.path.isfile(a):
-                files.append(a)
-            else:
-                raise FileNotFoundError(f"{p} does not exist or is not a valid file/directory")
-
-        images = []
-        for f in files:
-            suffix = f.split(".")[-1].lower()
-            if suffix in IMG_FORMATS:
-                images.append(f)
-        ni = len(images)
-
-        self.files = images
-        self.nf = ni  # number of files
-        self.ni = ni  # number of images
-        self.mode = "image"
-        self.vid_stride = vid_stride
-        self.bs = batch  # batch size
-        if self.nf == 0:
-            raise FileNotFoundError(f"No images or videos found in {p}")
-
-    def __iter__(self):
-        self.count = 0
-        return self
-
-    def __next__(self):
-        paths, imgs, info = [], [], []
-        while len(imgs) < self.bs:
-            if self.count >= self.nf:
-                if imgs:
-                    return paths, imgs, info
-                else:
-                    raise StopIteration
-
-            path = self.files[self.count]
-            im0 = imread(path)
-            if im0 is None:
-                logger.warning(f"WARNING ⚠️ Image Read Error {path}")
-            else:
-                paths.append(path)
-                imgs.append(im0)
-                info.append(f"image {self.count + 1}/{self.nf} {path}: ")
-            self.count += 1
-            if self.count >= self.ni:
-                break
-
-        return paths, imgs, info
-
-    def _new_video(self, path):
-        self.frame = 0
-        self.cap = cv2.VideoCapture(path)
-        self.fps = int(self.cap.get(cv2.CAP_PROP_FPS))
-        if not self.cap.isOpened():
-            raise FileNotFoundError(f"Failed to open video {path}")
-        self.frames = int(self.cap.get(cv2.CAP_PROP_FRAME_COUNT) / self.vid_stride)
-
-    def __len__(self):
-        return math.ceil(self.nf / self.bs)
-
-
-# Resize image with optional padding (letterboxing)
-def LetterBox(img, new_shape=(320, 320), auto=False, scaleFill=False, scaleup=True, center=True, stride=32):
-    shape = img.shape[:2]  # current shape (height, width)
-    if isinstance(new_shape, int):
-        new_shape = (new_shape, new_shape)
-
-    r = min(new_shape[0] / shape[0], new_shape[1] / shape[1])
-    if not scaleup:
-        r = min(r, 1.0)
-
-    new_unpad = int(round(shape[1] * r)), int(round(shape[0] * r))
-    dw, dh = new_shape[1] - new_unpad[0], new_shape[0] - new_unpad[1]
-    if auto:
-        dw, dh = np.mod(dw, stride), np.mod(dh, stride)
-
-    if center:
-        dw /= 2
-        dh /= 2
-
-    if shape[::-1] != new_unpad:
-        img = cv2.resize(img, new_unpad, interpolation=cv2.INTER_LINEAR)
-    top, bottom = int(round(dh - 0.1)) if center else 0, int(round(dh + 0.1))
-    left, right = int(round(dw - 0.1)) if center else 0, int(round(dw + 0.1))
-    img = cv2.copyMakeBorder(img, top, bottom, left, right, cv2.BORDER_CONSTANT, value=(114, 114, 114))
-
-    return img
-
-
-# Apply LetterBox to a batch of images
-def pre_transform(im, res):
-    return [LetterBox(img=x, new_shape=res) for x in im]
-
-
-# Image preprocessing: resize, normalize, and convert to tensor
-def preprocess(im, res):
-    device = "cpu"
-    not_tensor = not isinstance(im, torch.Tensor)
-    if not_tensor:
-        im = np.stack(pre_transform(im, res))
-        im = im[..., ::-1].transpose((0, 3, 1, 2))  # BGR to RGB, NHWC to NCHW
-        im = np.ascontiguousarray(im)
-        im = torch.from_numpy(im)
-
-    im = im.half() if device != "cpu" else im.float()
-    if not_tensor:
-        im /= 255  # normalize to [0, 1]
-    return im
-
-
-# Return an empty tensor or array of same shape and type
-def empty_like(x):
-    return (
-        torch.empty_like(x, dtype=torch.float32) if isinstance(x, torch.Tensor) else np.empty_like(x, dtype=np.float32)
-    )
-
-
-# Convert bounding boxes from (x_center, y_center, width, height) to (x1, y1, x2, y2)
-def xywh2xyxy(x):
-    assert x.shape[-1] == 4, f"input shape last dimension expected 4 but input shape is {x.shape}"
-    y = empty_like(x)
-    xy = x[..., :2]
-    wh = x[..., 2:] / 2
-    y[..., :2] = xy - wh
-    y[..., 2:] = xy + wh
-    return y
-
-
-# Perform non-maximum suppression (NMS) on predictions
-def non_max_suppression(
-    prediction,
-    conf_thres=0.25,
-    iou_thres=0.45,
-    classes=None,
-    agnostic=False,
-    multi_label=False,
-    labels=(),
-    max_det=300,
-    nc=0,
-    max_time_img=0.05,
-    max_nms=30000,
-    max_wh=7680,
-    in_place=True,
-    rotated=False,
-):
-    CONF_IDX = 4  # confidence score index
-    CLASS_IDX = 5  # class index
-    assert 0 <= conf_thres <= 1, f"Invalid Confidence threshold {conf_thres}, valid values are between 0.0 and 1.0"
-    assert 0 <= iou_thres <= 1, f"Invalid IoU {iou_thres}, valid values are between 0.0 and 1.0"
-
-    if isinstance(prediction, (list, tuple)):
-        prediction = prediction[0]
-    if classes is not None:
-        classes = torch.tensor(classes, device=prediction.device)
-
-    # Short-circuit path: filter predictions using confidence and class index
-    if prediction.shape[-1] == 6:
-        output = [pred[pred[:, CONF_IDX] > conf_thres][:max_det] for pred in prediction]
-        if classes is not None:
-            output = [pred[(pred[:, CLASS_IDX : CLASS_IDX + 1] == classes).any(1)] for pred in output]
-        return output
-
-    # Extended NMS logic
-    bs = prediction.shape[0]
-    nc = nc or (prediction.shape[1] - 4)
-    nm = prediction.shape[1] - nc - 4
-    mi = 4 + nc
-    xc = prediction[:, 4:mi].amax(1) > conf_thres
-
-    time_limit = 2.0 + max_time_img * bs
-    multi_label &= nc > 1
-
-    prediction = prediction.transpose(-1, -2)
-    if not rotated:
-        if in_place:
-            prediction[..., :4] = xywh2xyxy(prediction[..., :4])
-        else:
-            prediction = torch.cat((xywh2xyxy(prediction[..., :4]), prediction[..., 4:]), dim=-1)
-
-    t = time.time()
-    output = [torch.zeros((0, 6 + nm), device=prediction.device)] * bs
-    for xi, x in enumerate(prediction):
-        x = x[xc[xi]]
-        if not x.shape[0]:
-            continue
-
-        box, cls, mask = x.split((4, nc, nm), 1)
-
-        if multi_label:
-            i, j = torch.where(cls > conf_thres)
-            x = torch.cat((box[i], x[i, 4 + j, None], j[:, None].float(), mask[i]), 1)
-        else:
-            conf, j = cls.max(1, keepdim=True)
-            x = torch.cat((box, conf, j.float(), mask), 1)[conf.view(-1) > conf_thres]
-
-        if classes is not None:
-            x = x[(x[:, CLASS_IDX : CLASS_IDX + 1] == classes).any(1)]
-
-        n = x.shape[0]
-        if not n:
-            continue
-        if n > max_nms:
-            x = x[x[:, CONF_IDX].argsort(descending=True)[:max_nms]]
-
-        c = x[:, CLASS_IDX : CLASS_IDX + 1] * (0 if agnostic else max_wh)
-        scores = x[:, CONF_IDX]
-        boxes = x[:, :4] + c  # class offset for NMS across classes
-        i = torchvision.ops.nms(boxes, scores, iou_thres)
-        i = i[:max_det]
-        output[xi] = x[i]
-        if (time.time() - t) > time_limit:
-            logger.warning(f"WARNING ⚠️ NMS time limit {time_limit:.3f}s exceeded")
-            break
-
-    return output
-
-
-# Extract boxes, confidence, and class IDs from predictions
-def Boxes(data):
-    return {"xyxy": data[:, :4], "conf": data[:, -2], "cls": data[:, -1]}
-
-
-# Structure final detection results
-def Results(orig_img, path, names, boxes):
-    return {"orig_img": orig_img, "path": path, "names": names, "boxes": Boxes(boxes)}
-
-
-# Clip boxes to image boundaries
-def clip_boxes(boxes, shape):
-    if isinstance(boxes, torch.Tensor):
-        boxes[..., 0] = boxes[..., 0].clamp(0, shape[1])
-        boxes[..., 1] = boxes[..., 1].clamp(0, shape[0])
-        boxes[..., 2] = boxes[..., 2].clamp(0, shape[1])
-        boxes[..., 3] = boxes[..., 3].clamp(0, shape[0])
-    else:
-        boxes[..., [0, 2]] = boxes[..., [0, 2]].clip(0, shape[1])
-        boxes[..., [1, 3]] = boxes[..., [1, 3]].clip(0, shape[0])
-    return boxes
-
-
-# Rescale boxes from resized image to original image size
-def scale_boxes(img1_shape, boxes, img0_shape, ratio_pad=None, padding=True, xywh=False):
-    if ratio_pad is None:
-        gain = min(img1_shape[0] / img0_shape[0], img1_shape[1] / img0_shape[1])
-        pad = (
-            round((img1_shape[1] - img0_shape[1] * gain) / 2 - 0.1),
-            round((img1_shape[0] - img0_shape[0] * gain) / 2 - 0.1),
-        )
-    else:
-        gain = ratio_pad[0][0]
-        pad = ratio_pad[1]
-
-    if padding:
-        boxes[..., 0] -= pad[0]
-        boxes[..., 1] -= pad[1]
-        if not xywh:
-            boxes[..., 2] -= pad[0]
-            boxes[..., 3] -= pad[1]
-    boxes[..., :4] /= gain
-    return clip_boxes(boxes, img0_shape)
-
-
-# Postprocessing pipeline for object detection
-def postprocess(preds, img, orig_imgs, batch, names):
-    args = {"conf": 0.25, "iou": 0.7, "agnostic_nms": False, "max_det": 300, "classes": None}
-
-    preds = non_max_suppression(
-        preds,
-        args["conf"],
-        args["iou"],
-        agnostic=args["agnostic_nms"],
-        max_det=args["max_det"],
-        classes=args["classes"],
-    )
-
-    results = []
-    for pred, orig_img, img_path in zip(preds, orig_imgs, batch[0]):
-        pred[:, :4] = scale_boxes(img.shape[2:], pred[:, :4], orig_img.shape)
-        results.append(Results(orig_img, path=img_path, names=names, boxes=pred))
-
-    return results
-
-
-# Download COCO class names (with local fallback)
-def load_coco_class_names():
-    url = "https://raw.githubusercontent.com/pjreddie/darknet/master/data/coco.names"
-    path = f"models/demos/yolov4/demo/coco.names"
-    try:
-        response = requests.get(url, timeout=5)
-        if response.status_code == 200:
-            return response.text.strip().split("\n")
-    except requests.RequestException:
-        pass
-    if os.path.exists(path):
-        with open(path, "r") as f:
-            return [line.strip() for line in f.readlines()]
-
-    raise Exception("Failed to fetch COCO class names from both online and local sources.")
-
-
-def get_mesh_mappers(device):
-    if device.get_num_devices() > 1:
-        inputs_mesh_mapper = ttnn.ShardTensorToMesh(device, dim=0)
-        weights_mesh_mapper = ttnn.ReplicateTensorToMesh(device)
-        output_mesh_composer = ttnn.ConcatMeshToTensor(device, dim=0)
-    else:
-        inputs_mesh_mapper = None
-        weights_mesh_mapper = None
-        output_mesh_composer = None
-    return inputs_mesh_mapper, weights_mesh_mapper, output_mesh_composer
