+++ models/demos/yolov8x/tests/perf/test_e2e_performant.py
@@ -10,9 +10,9 @@ import torch
 from loguru import logger
 
 import ttnn
+from models.demos.utils.common_demo_utils import get_mesh_mappers
 from models.demos.yolov8x.common import YOLOV8X_L1_SMALL_SIZE
 from models.demos.yolov8x.runner.performant_runner import YOLOv8xPerformantRunner
-from models.demos.yolov8x.tt.ttnn_yolov8x_utils import get_mesh_mappers
 from models.utility_functions import run_for_wormhole_b0
