+++ models/demos/yolov9c/demo/demo_utils.py
@@ -10,132 +10,11 @@ from pathlib import Path
 import cv2
 import matplotlib.pyplot as plt
 import numpy as np
-import requests
 import torch
 import torch.nn.functional as F
 from loguru import logger
-from ultralytics import YOLO
 
-from models.demos.yolov9c.reference.yolov9c import YoloV9
-from models.experimental.yolo_eval.utils import non_max_suppression, scale_boxes
-
-
-def imread(filename: str, flags: int = cv2.IMREAD_COLOR):
-    return cv2.imdecode(np.fromfile(filename, np.uint8), flags)
-
-
-IMG_FORMATS = {"bmp", "dng", "jpeg", "jpg", "mpo", "png", "tif", "tiff", "webp", "pfm", "heic"}
-
-
-class LoadImages:
-    def __init__(self, path, batch=1, vid_stride=1):
-        files = []
-        for p in sorted(path) if isinstance(path, (list, tuple)) else [path]:
-            a = str(Path(p).absolute())
-            if os.path.isdir(a):
-                for f in os.listdir(a):
-                    if f.lower().endswith((".jpg", ".jpeg", ".png", ".bmp")):
-                        files.append(os.path.join(a, f))
-            elif os.path.isfile(a):
-                files.append(a)
-            else:
-                raise FileNotFoundError(f"{p} does not exist or is not a valid file/directory")
-
-        images = []
-        for f in files:
-            suffix = f.split(".")[-1].lower()
-            if suffix in IMG_FORMATS:
-                images.append(f)
-        ni = len(images)
-        self.files = images
-        self.nf = ni
-        self.ni = ni
-        self.mode = "image"
-        self.vid_stride = vid_stride
-        self.bs = batch
-        if self.nf == 0:
-            raise FileNotFoundError(f"No images or videos found in {p}")
-
-    def __iter__(self):
-        self.count = 0
-        return self
-
-    def __next__(self):
-        paths, imgs, info = [], [], []
-        while len(imgs) < self.bs:
-            if self.count >= self.nf:
-                if imgs:
-                    return paths, imgs, info
-                else:
-                    raise StopIteration
-
-            path = self.files[self.count]
-            im0 = imread(path)
-            if im0 is None:
-                logger.warning(f"WARNING ⚠️ Image Read Error {path}")
-            else:
-                paths.append(path)
-                imgs.append(im0)
-                info.append(f"image {self.count + 1}/{self.nf} {path}: ")
-            self.count += 1
-            if self.count >= self.ni:
-                break
-
-        return paths, imgs, info
-
-    def _new_video(self, path):
-        self.frame = 0
-        self.cap = cv2.VideoCapture(path)
-        self.fps = int(self.cap.get(cv2.CAP_PROP_FPS))
-        if not self.cap.isOpened():
-            raise FileNotFoundError(f"Failed to open video {path}")
-        self.frames = int(self.cap.get(cv2.CAP_PROP_FRAME_COUNT) / self.vid_stride)
-
-    def __len__(self):
-        return math.ceil(self.nf / self.bs)
-
-
-def load_coco_class_names():
-    url = "https://raw.githubusercontent.com/pjreddie/darknet/master/data/coco.names"
-    path = f"models/demos/yolov4/demo/coco.names"
-    response = requests.get(url)
-    try:
-        response = requests.get(url, timeout=5)
-        if response.status_code == 200:
-            return response.text.strip().split("\n")
-    except requests.RequestException:
-        pass
-    if os.path.exists(path):
-        with open(path, "r") as f:
-            return [line.strip() for line in f.readlines()]
-
-    raise Exception("Failed to fetch COCO class names from both online and local sources.")
-
-
-def load_torch_model(use_weights_from_ultralytics=True, module=None, model_task="segment"):
-    state_dict = None
-
-    weights = "yolov9c-seg.pt" if model_task == "segment" else "yolov9c.pt"
-    enable_segment = model_task == "segment"
-
-    if use_weights_from_ultralytics:
-        torch_model = YOLO(weights)  # Use "yolov9c.pt" weight for detection
-        torch_model.eval()
-        state_dict = torch_model.state_dict()
-
-    model = YoloV9(enable_segment=enable_segment)
-    state_dict = model.state_dict() if state_dict is None else state_dict
-
-    ds_state_dict = {k: v for k, v in state_dict.items()}
-    new_state_dict = {}
-    for (name1, parameter1), (name2, parameter2) in zip(model.state_dict().items(), ds_state_dict.items()):
-        if isinstance(parameter2, torch.FloatTensor):
-            new_state_dict[name1] = parameter2
-
-    model.load_state_dict(new_state_dict)
-    model.eval()
-
-    return model
+from models.demos.utils.common_demo_utils import non_max_suppression, scale_boxes
 
 
 def get_consistent_color(index):
@@ -170,16 +49,6 @@ def save_seg_predictions_by_model(result, save_dir, image_path, model_name):
     out_path = os.path.join(save_dir, model_name, f"segmentation_{timestamp}.jpg")
     cv2.imwrite(out_path, overlay_bgr)
     logger.info(f"Saved to {out_path}")
-
-
-# For Segmentation task
-
-
-def imread(filename: str, flags: int = cv2.IMREAD_COLOR):
-    return cv2.imdecode(np.fromfile(filename, np.uint8), flags)
-
-
-IMG_FORMATS = {"bmp", "dng", "jpeg", "jpg", "mpo", "png", "tif", "tiff", "webp", "pfm", "heic"}
 
 
 class LetterBox:
